bug: fixed multiple instances of place eventManager
