+++ scripts/main.js
@@ -1,5 +1,3 @@
-import { parkList } from "./park/parkList.js";
 import { parkSearch } from "./park/parkSearch.js";
 
-parkList();
 parkSearch();
