+++ scripts/park/park.js
@@ -1,32 +1,7 @@
-import { useParks, setParkCode } from "./parkProvider.js";
-import { placeList } from "../place/placeList.js";
+import { setParkCode } from "./parkProvider.js";
 
 export const parkComponent = (park) => {
-  eventManager();
   return render(park);
-};
-
-const eventManager = () => {
-  const eventHub = document.querySelector(".container");
-
-  eventHub.addEventListener("click", (clickEvent) => {
-    if (clickEvent.target.className === "park-button") {
-      clickEvent.preventDefault();
-
-      let parks = useParks();
-
-      const selectedPark = parks.find(
-        (park) => park.id === clickEvent.target.id
-      );
-
-      document.querySelector(".park-list").innerHTML =
-        renderDetail(selectedPark);
-
-      // run scripts for attraction/places lists here. Or maybe use events?
-      // This may be problematic since it is running for each instance in the park list.
-      placeList();
-    }
-  });
 };
 
 const render = (park) => {
@@ -35,12 +10,12 @@ const render = (park) => {
       <h2 class="park-name">${park.fullName}</h2>
       <image class="park-image" src="${park.images[0].url}" alt="${park.images[0].altText}">
       <p class="park-description">${park.description}</p>
-      <button class="park-button" id=${park.id}>Plan a Trip</button>
+      <button class="park-button-${park.id}" id=${park.id}>Plan a Trip</button>
     </div>
     `;
 };
 
-const renderDetail = (park) => {
+export const renderDetail = (park) => {
   setParkCode(park.parkCode);
 
   // Going to add more detail later. Leaving simple for now.
