+++ scripts/park/parkList.js
@@ -1,10 +1,34 @@
 import { useParks } from "./parkProvider.js";
-import { parkComponent } from "./park.js";
+import { parkComponent, renderDetail } from "./park.js";
+import { placeList } from "../place/placeList.js";
 
 export const parkList = () => {
   let parks = useParks();
 
+  eventManager();
   render(parks);
+};
+
+const eventManager = () => {
+  const eventHub = document.querySelector(".container");
+
+  eventHub.addEventListener("click", (clickEvent) => {
+    if (clickEvent.target.className === `park-button-${clickEvent.target.id}`) {
+      clickEvent.preventDefault();
+
+      let parks = useParks();
+
+      const selectedPark = parks.find(
+        (park) => park.id === clickEvent.target.id
+      );
+
+      document.querySelector(".park-list").innerHTML =
+        renderDetail(selectedPark);
+
+      // run scripts for attraction/places lists here. Or maybe use events?
+      placeList();
+    }
+  });
 };
 
 const render = (parks) => {
