+++ scripts/place/place.js
@@ -2,7 +2,7 @@ export const placeComponent = (place) => {
   return `
   <h3>${place.title}</h3>
   <image class="place-image" src="${place.images[0].url}">
-  <div class="place-bodyw">
+  <div class="place-body">
     ${place.bodyText}
   </div>
   <label for ="${place.id}">Add to trip</label>
